move certain states to MCPServerSelector
